ChatAPI/requirements.txt
openai>=0.27.0
requests>=2.28.2
Flask>=2.0.2
Flask-JWT-Extended>=4.4.4
flask_cors>=3.0.10
pandas>=1.4.2
uwsgi
flask_sslify
pillow
googletrans
flask_gzip
mtranslate